requirements.txt
request
urllib3[secure]
click